+++ VisuRDF/visurdfparametreur.h
@@ -38,9 +38,9 @@ public:
     int getFontSize();
     string getParamMode();
     map<int,QBrush*> getListePinceau();
-    int getPourcentagePolice();
-    int getPourcentagePoliceHauteur();
-    int getTailleMax();
+    float getPourcentagePolice();
+    float getPourcentagePoliceHauteur();
+    float getTailleMax();
 
     void setMode(string mode);
     void setPourcentagePolice(int p);
@@ -60,9 +60,9 @@ private:
     int interligne;
     string mode;
     int coloration;
-    int pourcentagePoliceHauteur;
-    int pourcentagePolice;
-    int tailleMax;
+    float pourcentagePoliceHauteur;
+    float pourcentagePolice;
+    float tailleMax;
 };
